I am wondering whether using json scripting control the crawling or using native java crawling.
scripting will need a script parser, which need a lot of code to work
with the parser.
hard, but it may be more universal, if i can write it perfectly.

by using native java code, i need to write a lot of crawler, but it's
easy. however, i need to update code when remote changed.

now, i decide to use java native code. but this will be committed as
history.