I become hisitate of the three ways of analyze the html.
1. HttpClient + RegularExpression, the core expression may be a little
hard, but fit in general propose.
2. HttpClient + HtmlCleaner, which transform the html to formed xml, and
using XPath to analyze the document. This is the industry standard of
html paresing.
3. Jsoup, a lib focusing on analyze the html like jQuery, which is
simple.

Originally, I decided to the the 2nd way, but it seem the 3rd is much
more simple.

Maybe I will implements all the three ways to see which is better. But
now, I decided using the 3rd way currently.

Signed-off-by: Winguse <github@winguse.com>